Add authentication and validation checks, plus misc refactoring

+++ src/main/java/com/revaturemax/services/NotesService.java
@@ -1,7 +1,9 @@
 package com.revaturemax.services;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.revaturemax.models.Employee;
 import com.revaturemax.models.Notes;
+import com.revaturemax.models.Token;
 import com.revaturemax.repositories.NotesRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,36 +15,24 @@ public class NotesService {
 
     @Autowired
     private ObjectMapper objectMapper;
-
     @Autowired
     private NotesRepository notesRepository;
 
-    public ResponseEntity<String> setNotes(long employeeId, Notes notes) {
+    public ResponseEntity<String> setNotes(Token token, long employeeId, Notes notes) {
+        if (token.getEmployeeId() != employeeId) return new ResponseEntity<String>(HttpStatus.FORBIDDEN);
         validateNotes(notes);
-        //long employeeId = pulled/passed from JWT
-        //notes.setEmployee(new Employee(employeeId));
+        notes.setEmployee(new Employee(employeeId));
         if (notes.getNotes().equals("")) {
             notesRepository.delete(notes);
-            return new ResponseEntity<String>(HttpStatus.OK);
+            return new ResponseEntity<>(HttpStatus.OK);
         } else {
             notes = notesRepository.save(notes);
-            return new ResponseEntity<String>(notes.getId().toString(), HttpStatus.OK);
+            return new ResponseEntity<>(notes.getId().toString(), HttpStatus.OK);
         }
     }
 
     private void validateNotes(Notes notes) {
         //check topic set
-    }
-
-    public void deleteNotes(long id) { notesRepository.deleteById(id);
-    }
-
-   public Notes updateNotes(Long id, Notes newNotes){
-       Notes updateN = notesRepository.getOne(id);
-       updateN.setNotes(newNotes.getNotes());
-//       updateN.setTopic(newNotes.getTopic());
-       notesRepository.save(updateN);
-       return updateN;
     }
 
 }
